requirements.txt
-i https://pypi.org/simple
aiohappyeyeballs==2.4.3; python_version >= '3.8'
aiohttp==3.10.9; python_version >= '3.8'
aiosignal==1.3.1; python_version >= '3.7'
attrs==24.2.0; python_version >= '3.7'
cachetools==5.5.0; python_version >= '3.7'
certifi==2024.8.30; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
discord.py==2.4.0; python_version >= '3.8'
freezegun==1.5.1; python_version >= '3.7'
frozenlist==1.4.1; python_version >= '3.8'
google-api-core==2.20.0; python_version >= '3.7'
google-api-python-client==2.147.0; python_version >= '3.7'
google-auth==2.35.0; python_version >= '3.7'
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1; python_version >= '3.6'
googleapis-common-protos==1.65.0; python_version >= '3.7'
httplib2==0.22.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.10; python_version >= '3.6'
iniconfig==2.0.0; python_version >= '3.7'
multidict==6.1.0; python_version >= '3.8'
oauthlib==3.2.2; python_version >= '3.6'
packaging==24.1; python_version >= '3.8'
pluggy==1.5.0; python_version >= '3.8'
proto-plus==1.24.0; python_version >= '3.7'
protobuf==5.28.2; python_version >= '3.8'
pyasn1==0.6.1; python_version >= '3.8'
pyasn1-modules==0.4.1; python_version >= '3.8'
pyparsing==3.1.4; python_version >= '3.1'
pytest==8.3.3; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
requests-oauthlib==2.0.0; python_version >= '3.4'
rsa==4.9; python_version >= '3.6' and python_version < '4'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.2.3; python_version >= '3.8'
yarl==1.13.1; python_version >= '3.8'
